requirements.txt
openai>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
python-dotenv>=1.0.0
replicate>=0.15.0
Pillow>=10.0.0
pandas>=2.0.0  # NEW: For user management tables